Move i18n validation out of app metadata for fine granular hmr events
